Modified configs to allow test runner to work inside IDE.

+++ sct-integration-test-mmb/sct-integration-tests/src/test/java/org/ihtsdo/sct/RegressionRetrieveStatedTest.java
@@ -123,18 +123,18 @@ public class RegressionRetrieveStatedTest {
 
         cEditPathBI = snomedPath;
         cEditPathNid = cEditPathBI.getConceptNid();
-        cEditPosSet = new PositionSetReadOnly(tf.newPosition(cEditPathBI, Integer.MAX_VALUE));
+        cEditPosSet = new PositionSetReadOnly(tf.newPosition(cEditPathBI, Long.MAX_VALUE));
 
         cEditPathListPositionBI = new ArrayList<PositionBI>();
-        cEditPathListPositionBI.add(tf.newPosition(cEditPathBI, Integer.MAX_VALUE));
+        cEditPathListPositionBI.add(tf.newPosition(cEditPathBI, Long.MAX_VALUE));
         getPathOrigins(cEditPathListPositionBI, cEditPathBI);
 
         cViewPathBI = snomedPath;
         cViewPathNid = cViewPathBI.getConceptNid();
-        cViewPosSet = new PositionSetReadOnly(tf.newPosition(cViewPathBI, Integer.MAX_VALUE));
+        cViewPosSet = new PositionSetReadOnly(tf.newPosition(cViewPathBI, Long.MAX_VALUE));
 
         cViewPathListPositionBI = new ArrayList<PositionBI>();
-        cViewPathListPositionBI.add(tf.newPosition(cViewPathBI, Integer.MAX_VALUE));
+        cViewPathListPositionBI.add(tf.newPosition(cViewPathBI, Long.MAX_VALUE));
         getPathOrigins(cViewPathListPositionBI, cViewPathBI);
     }
